api/requirements.txt
Django==2.2.18
django-cleanup==5.1.0
django-cors-headers==3.7.0
django-countries==7.0
django-filter==2.4.0
django-hashid-field==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
hashids==1.3.1
numpy==1.20.2
Pillow==8.1.0
psycopg2==2.8.6
PyJWT==2.0.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tinytag==1.5.0
whitenoise==5.2.0
